cd: Move rel classes to relational.py, improve docstrs+test
